Debasish616/HackIndia/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
azure-communication-email
